% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_decompress_pbix.R
\name{f_decompress_pbix}
\alias{f_decompress_pbix}
\title{Decompress '.pbix' to a Collection of Files}
\usage{
f_decompress_pbix(input_file_pbix, collection_files_pbix)
}
\arguments{
\item{input_file_pbix}{Path of the input '.pbix'.}

\item{collection_files_pbix}{Directory of the collection of files.}
}
\value{
None
}
\description{
A '.pbix' is decompressed, making its collection of files
available for manipulation.
}
\examples{
\dontrun{
# Get dummy data ------------------------------------------------------------
# Create a temporary directory
temp_dir <- file.path(tempdir(),"functionTest")
if(!dir.exists(temp_dir)) {
	dir.create(temp_dir)
}
sample_file_name <- "OR_sample_func.pbix"
pathFileSample <- file.path(temp_dir, sample_file_name)

# See if dummy data already exists in temporary directory
parent_temp_dir <- dirname(temp_dir)
existing_file <- list.files(parent_temp_dir,
pattern = sample_file_name, recursive = TRUE, full.names = TRUE)

# Download the sample .pbix if it doesn't exist
if (length(existing_file) == 0) {
   url_pt1 <- "https://github.com/KoenVerbeeck/PowerBI-Course/blob/"
   url_pt2 <- "master/pbix/TopMovies.pbix?raw=true"
   url <- paste0(url_pt1, url_pt2)
   req <- download.file(url, destfile = pathFileSample, mode = "wb")
} else {
   pathFileSample <- existing_file[1]
}
# Do stuff ------------------------------------------------------------------

output_pbix_file <- gsub("OR_sample_func.pbix", "OR_unzip_pbix",
pathFileSample)
# Run the function
f_decompress_pbix(pathFileSample, output_pbix_file)
  }
}
\author{
Don Diproto
}
